refactor(refund-tables): update refund table data structure and remove unused components
