+++ apps/web/src/app/[lang]/(main)/(unirefund)/settings/templates/refund-tables/[id]/form.tsx
@@ -1,16 +1,26 @@
 "use client";
 import { Button } from "@/components/ui/button";
 import { toast } from "@/components/ui/sonner";
-import type { UniRefund_ContractService_Refunds_RefundTableHeaders_RefundTableHeaderDto } from "@ayasofyazilim/saas/ContractService";
+import type {
+  UniRefund_ContractService_Refunds_RefundTableDetails_RefundTableDetailCreateDto,
+  UniRefund_ContractService_Refunds_RefundTableHeaders_RefundTableHeaderDto,
+} from "@ayasofyazilim/saas/ContractService";
 import { $UniRefund_ContractService_Refunds_RefundTableHeaders_RefundTableHeaderUpdateDto } from "@ayasofyazilim/saas/ContractService";
 import ConfirmDialog from "@repo/ayasofyazilim-ui/molecules/confirm-dialog";
+import { tanstackTableEditableColumnsByRowData } from "@repo/ayasofyazilim-ui/molecules/tanstack-table/utils";
 import { SchemaForm } from "@repo/ayasofyazilim-ui/organisms/schema-form";
+import { TableField } from "@repo/ayasofyazilim-ui/organisms/schema-form/fields";
 import { createUiSchemaWithResource } from "@repo/ayasofyazilim-ui/organisms/schema-form/utils";
 import { ActionList } from "@repo/ui/action-button";
+import { PlusCircle, Trash2 } from "lucide-react";
 import { useRouter } from "next/navigation";
 import { handlePutResponse } from "src/actions/core/api-utils-client";
-import { putRefundTableHeadersApi } from "src/actions/unirefund/ContractService/put-actions";
+import { putRefundTableHeadersByIdApi } from "src/actions/unirefund/ContractService/put-actions";
 import type { ContractServiceResource } from "src/language-data/unirefund/ContractService";
+
+type TypeWithId<Type, IdType = string> = Type & {
+  id: IdType;
+};
 
 function Form({
   response,
@@ -29,17 +39,54 @@ function Form({
       name: {
         "ui:className": "md:col-span-full",
       },
+      refundTableDetails: {
+        "ui:className": "md:col-span-full",
+        "ui:field": "RebateTable",
+      },
     },
+  });
+  const RebateTableColumns = tanstackTableEditableColumnsByRowData<
+    TypeWithId<UniRefund_ContractService_Refunds_RefundTableDetails_RefundTableDetailCreateDto>
+  >({
+    rows: $UniRefund_ContractService_Refunds_RefundTableHeaders_RefundTableHeaderUpdateDto
+      .properties.refundTableDetails.items.properties,
+    excludeColumns: ["extraProperties"],
   });
   return (
     <SchemaForm
+      fields={{
+        RebateTable: TableField<
+          TypeWithId<UniRefund_ContractService_Refunds_RefundTableDetails_RefundTableDetailCreateDto>
+        >({
+          editable: true,
+          columns: RebateTableColumns,
+          data: response.refundTableDetails || [],
+          fillerColumn: "id",
+          tableActions: [
+            {
+              type: "create-row",
+              actionLocation: "table",
+              cta: languageData["RefundTables.Form.RefundTableDetails.add"],
+              icon: PlusCircle,
+            },
+          ],
+          rowActions: [
+            {
+              actionLocation: "row",
+              cta: languageData["Rebate.Form.rebateTableDetails.delete"],
+              icon: Trash2,
+              type: "delete-row",
+            },
+          ],
+        }),
+      }}
       formData={response}
       onSubmit={(data) => {
         const formData = {
           id: response.id,
           requestBody: data.formData,
         };
-        void putRefundTableHeadersApi(formData).then((res) => {
+        void putRefundTableHeadersByIdApi(formData).then((res) => {
           handlePutResponse(res, router);
         });
       }}
@@ -48,9 +95,8 @@ function Form({
       }
       uiSchema={uiSchema}
       useDefaultSubmit={false}
-      withScrollArea={false}
     >
-      <ActionList className="mb-4 border-0 p-0">
+      <ActionList className="sticky bottom-0 z-10 mb-4 border-0 bg-white p-0 py-4">
         <ConfirmDialog
           confirmProps={{
             onConfirm: () => {
